[refactor] presentation - 공통요소 분리
