+++ presentation/src/main/AndroidManifest.xml
@@ -2,7 +2,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
     <application>
         <activity
-            android:name=".MainActivity"
+            android:name=".common.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.ToucheeseClean">
             <intent-filter>
