+++ presentation/src/main/java/com/toucheese/presentation/common/navigation/Navigation.kt
@@ -1,0 +1,22 @@
+package com.toucheese.presentation.common.navigation
+
+import androidx.navigation.NavController
+
+class Navigation {
+
+    companion object {
+        fun bottomNavClicked(selectedTab: Int, navController: NavController) {
+            when (selectedTab) {
+                // 홈화면으로 이동
+                0 -> {
+                    navController.navigate(Screen.Home.route) {
+                        popUpTo(navController.graph.id) {
+                            inclusive = true
+                        }
+                    }
+
+                }
+            }
+        }
+    }
+}
