+++ presentation/src/main/java/com/toucheese/presentation/common/utils/Concept.kt
@@ -1,0 +1,18 @@
+package com.toucheese.presentation.common.utils
+
+import com.toucheese.presentation.R
+
+class Concept {
+
+    companion object {
+
+        val conceptArray = arrayOf(
+            Pair(R.drawable.image_concept1, R.string.concept_image1),
+            Pair(R.drawable.image_concept2, R.string.concept_image2),
+            Pair(R.drawable.image_concept3, R.string.concept_image3),
+            Pair(R.drawable.image_concept4, R.string.concept_image4),
+            Pair(R.drawable.image_concept5, R.string.concept_image5),
+            Pair(R.drawable.image_concept6, R.string.concept_image6),
+        )
+    }
+}
