+++ presentation/src/main/java/com/toucheese/presentation/ui/screens/SignUpAdditionalInfoScreen.kt
@@ -35,9 +35,9 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.toucheese.presentation.R
-import com.toucheese.presentation.ui.component.appbar.BasicTopAppbarComponent
-import com.toucheese.presentation.ui.component.button.ButtonComponent
-import com.toucheese.presentation.ui.component.textfield.OutlinedTextFieldErrorComponent
+import com.toucheese.presentation.common.component.appbar.BasicTopAppbarComponent
+import com.toucheese.presentation.common.component.button.ButtonComponent
+import com.toucheese.presentation.common.component.textfield.OutlinedTextFieldErrorComponent
 import com.toucheese.presentation.ui.viewmodel.SignUpViewModel
 import kotlinx.coroutines.launch
