--- presentation/src/main/java/com/toucheese/presentation/utils/Screen.kt
@@ -1,16 +1,0 @@
-package com.toucheese.presentation.utils
-
-sealed class Screen(val route: String) {
-    // Home
-    data object Home: Screen("Home")
-    // Sign In
-    data object SignIn: Screen("SignIn")
-    // Studio List
-    data object StudioFilter: Screen("StudioFilter/{conceptId}")
-    // Additional info
-    data object AdditionalInfo: Screen("AdditionalInfo")
-    // Sign Up
-    data object SignUp: Screen("SignUp")
-    // Sign Up Additional Info
-    data object SignUpAdditionalInfo: Screen("SignUpAdditionalInfo/{email}/{password}")
-}
